[refactor] domain entity 및 service 로직 개선

+++ src/02-domain/service/article.comment.service.ts
@@ -1,5 +1,6 @@
 import { IArticleCommentService } from "../../01-inbound/port/services/i.article.comment.service";
-import { ArticleCommentRequest } from "../../01-inbound/request/req.validator";
+import { ArticleCommentDto } from "../../01-inbound/request/req.validator";
+import { ArticleCommentResDto } from "../../01-inbound/response/article.comment.response";
 import { Authenticator } from "../../external/authenticator";
 import { IBaseRepository } from "../port/I.base.repository";
 
@@ -17,33 +18,50 @@ export class ArticleCommentService implements IArticleCommentService {
 
 
 
-    async createArticleComment(dto: ArticleCommentRequest) {
+    async createArticleComment(dto: ArticleCommentDto) {
         const { articleId, userId, content } = dto;
         const articleCommentResDto = await this.#repos.articleComment.save(userId, articleId, content);
         await this.#repos.notification.createArticleCommentNotification(userId);
-        return articleCommentResDto;
+        return new ArticleCommentResDto(articleCommentResDto);   
 
     }
 
     async getArticleComments(articleId: string) {
         const articleCommentResDtos = await this.#repos.articleComment.findArticleComments(articleId);
-        return articleCommentResDtos;
+        return articleCommentResDtos.map(dto => new ArticleCommentResDto(dto));
     }
 
-    async deleteArticleComments(commentId: string) {
-        await this.#repos.articleComment.deleteArticleComment(commentId);
-    }
-
-    async updateArticleComment(dto: ArticleCommentRequest) {
+    async updateArticleComment(dto: ArticleCommentDto) {
         const { articleId, commentId, content, userId } = dto;
         if (!commentId) {
             throw new Error('Comment ID is required for updating a comment.');
         }
 
 
-        const articleCommentResDto = await this.#repos.articleComment.update(userId, articleId, commentId, content);
+        const articleComment = await this.#repos.articleComment.findArticleComment(commentId);
+        if (!articleComment) {
+            throw new Error('Article comments not found.');
+        }
+        if (articleComment.userId !== userId) {
+            throw new Error('Unauthorized to update this comment.');
+        }
 
-        return articleCommentResDto;
+        articleComment.update(content);
+        const articleCommentResDto = await this.#repos.articleComment.update(articleComment);
+        return new ArticleCommentResDto(articleCommentResDto);
+    }
+
+    async deleteArticleComments(articleId: string, commentId: string, userId: string) {
+        const articleComment = await this.#repos.articleComment.findArticleComment(commentId);
+        if (!articleComment) {
+            throw new Error('Article comments not found.');
+        }
+        
+        if (articleComment.userId !== userId) {
+            throw new Error('Unauthorized to delete this comment.');
+        }
+
+        await this.#repos.articleComment.deleteArticleComment(commentId);
     }
 
 }
